Move MonthInput classes to monthinput package
