+++ docs/com/checkout/android_sdk/Input/MonthInput.Months.html
@@ -102,7 +102,7 @@ var activeTableTab = "activeTableTab";
 <li>java.lang.Enum&lt;<a href="../../../../com/checkout/android_sdk/Input/MonthInput.Months.html" title="enum in com.checkout.sdk.input">MonthInput.Months</a>&gt;</li>
 <li>
 <ul class="inheritance">
-<li>com.checkout.sdk.input.MonthInput.Months</li>
+<li>com.checkout.sdk.monthinput.MonthInput.Months</li>
 </ul>
 </li>
 </ul>
