+++ docs/com/checkout/android_sdk/Input/MonthInput.html
@@ -117,7 +117,7 @@ var activeTableTab = "activeTableTab";
 <li>android.support.v7.widget.AppCompatSpinner</li>
 <li>
 <ul class="inheritance">
-<li>com.checkout.sdk.input.MonthInput</li>
+<li>com.checkout.sdk.monthinput.MonthInput</li>
 </ul>
 </li>
 </ul>
@@ -391,7 +391,7 @@ extends android.support.v7.widget.AppCompatSpinner</pre>
 <div class="block">Populate the spinner with all the month of the year</div>
 </li>
 </ul>
-<a name="setMonthListener-com.checkout.sdk.input.MonthInput.MonthListener-">
+<a name="setMonthListener-com.checkout.sdk.monthinput.MonthInput.MonthListener-">
 <!--   -->
 </a>
 <ul class="blockListLast">
